Hoàn thiện quản lý giáo viên

+++ components/UI/List/ListItem.js
@@ -3,23 +3,60 @@ import { RiCharacterRecognitionFill } from "react-icons/ri";
 import { FaBirthdayCake } from "react-icons/fa";
 import Link from "next/link";
 import { useRouter } from "next/router";
+import { useDispatch } from "react-redux";
+import { LoadingActions } from "../../../store/redux/loading/loading-slice";
+
 const PhanTuDanhSach = (props) => {
   const router = useRouter();
+  const dispatchFn = useDispatch();
   //Các thông tin cần từ props
-  const { name, birthday, _id } = props;
+  const { name, birthday, _id, type } = props;
   // const viewDetaiHandler = (id) => {console.log('viewDetail')};
   const deleteItemHandler = (id) => {
-    //Định danh lạ id thăng mongodb mới hiểu
-    fetch("/api/quan-ly-hoc-sinh", {
-      method: "DELETE",
-      body: JSON.stringify(id),
-      headers: { "Content-Type": "application/json" },
-    });
-    //Xử lý trỏ về trang qlhs để load lại data rồi mới push lại trang hiện tại
-    router.replace('/quan-ly-hoc-sinh')
-    setTimeout(()=>{
-      router.replace(router.pathname)
-    },500)
+    dispatchFn(LoadingActions.activeLoading());
+    //Xử lý xóa học sinh
+    if (type === "student") {
+      //Định danh lạ id thăng mongodb mới hiểu
+      fetch("/api/quan-ly-hoc-sinh", {
+        method: "DELETE",
+        body: JSON.stringify(id),
+        headers: { "Content-Type": "application/json" },
+      })
+        .then((response) => {
+          dispatchFn(LoadingActions.deactiveLoading()),
+            router.replace("/quan-ly-hoc-sinh");
+          router.replace(router.pathname);
+        })
+        .catch((error) => {
+          dispatchFn(LoadingActions.deactiveLoading()),
+            router.replace("/quan-ly-hoc-sinh");
+          router.replace(router.pathname);
+        });
+    }
+
+    //Xử lý xóa giáo viên
+    if (type === "teacher") {
+      fetch("/api/quan-ly-giao-vien", {
+        method: "DELETE",
+        body: JSON.stringify(id),
+        headers: { "Content-Type": "application/json" },
+      })
+        .then((response) => {
+          dispatchFn(LoadingActions.deactiveLoading()),
+            router.replace("/quan-ly-giao-vien");
+          router.replace("/quan-ly-giao-vien/ds-giao-vien");
+        })
+        .catch((error) => {
+          dispatchFn(LoadingActions.deactiveLoading()),
+            router.replace("/quan-ly-giao-vien");
+          router.replace("/quan-ly-giao-vien/ds-giao-vien");
+        });
+      //Xử lý trỏ về trang qlgv load lại data ròi mới push lại trang hiện tại
+      // router.replace("/quan-ly-giao-vien");
+      // setTimeout(() => {
+      //   router.replace(router.pathname);
+      // }, 500);
+    }
   };
   return (
     <div className={classes.items}>
@@ -38,12 +75,14 @@ const PhanTuDanhSach = (props) => {
         </div>
       </div>
       <div className={classes.actions}>
-        <Link href={`/quan-ly-hoc-sinh/chi-tiet/${_id}`}>
-          <button
-            className="btn-sm btn-sm-detail"
-          >
-            Chi tiết
-          </button>
+        <Link
+          href={
+            type === "student"
+              ? `/quan-ly-hoc-sinh/chi-tiet/${_id}`
+              : `/quan-ly-giao-vien/chi-tiet/${_id}`
+          }
+        >
+          <button className="btn-sm btn-sm-detail">Chi tiết</button>
         </Link>
 
         <button
